requirements.txt
appdirs==1.4.2
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.1.5
django-allauth==0.39.1
django-autofixture==0.12.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-decouple==2.1
djangorestframework==3.9.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
packaging==16.8
Pillow==6.0.0
pipenv==2018.11.26
psycopg2==2.8.3
pylint==2.3.1
pyparsing==2.1.10
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.10.0
sqlparse==0.2.4
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.4.1
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
wrapt==1.11.1
